+++ wgpu-hal/examples/halmark/main.rs
@@ -357,7 +357,7 @@ impl<A: hal::Api> Example<A> {
             mipmap_filter: wgt::FilterMode::Nearest,
             lod_clamp: 0.0..32.0,
             compare: None,
-            anisotropy_clamp: 1.0,
+            anisotropy_clamp: 1,
             border_color: None,
         };
         let sampler = unsafe { device.create_sampler(&sampler_desc).unwrap() };